pcf8574: Set unused port to outputs
If the unused floating ports are left as inputs, they have no defined
state and may trigger a huge amount of state changes. This causes
increased power usage and spurious interrupts on the INT port.

The TI datasheet mentions in §9.2.D:
"P6 and P7 are not used and must be configured as outputs."

So all unused port must be set to 0. There is one slight complication: if
we start with all ports as outputs, they might flap during setup. To prevent
that, all ports start out as inputs during setup(). The final mode_mask is
only set and written to the device after all setup completed, ie during the
first loop().