[online_image] Last-Modified-Date and ETag response caching
Use the Last-Modified-Date and ETag response headers (if present) when making subsequent requests to implement conditional response caching. If the data on the server is not modified, then a 304 Not Modified response results and the online_image component will reuse the previously received data instead of downloading and processing it again.

See: https://developer.mozilla.org/en-US/docs/Web/HTTP/Caching