mlx90393: Add verification for register contents
After months of successful usage, the data suddenly changed.
After a reset it went back to normal. This is not good for long running sensors.

To fix this, a verification system is added.
It regularly checks all registers of the MLX90393.
If any value is found to be incorrect, the device is reset and all values are re-written.