Fix end_of_scan_ not being called while disconnecting
If at least one client is in disconnecting state,
we will not take the lock to figure out if the scanner
needs to be started or end of scan needs to
be declared. Since we do not stop the scanner
for disconnecting, and only connecting, we should
only avoid taking the lock if we are connecting.