If in continuous mode check sensor in loop()
Continuously check the sensor status in loop and then update whenever the sensor is ready if in periodic mode.

If in single shot mode, initiate a read in update and then report the result when the read finishes.

This should avoid losing data due to mistyped update() calls in periodic mode.