Sort resolved IP addresses for dashboard
In https://github.com/esphome/esphome/pull/7414 I fixed the name
resolution in helpers.resolve_ip_address() to return a *list* of
addresses sorted in an approximation of RFC6724 preference.

I missed the fact that the dashboard code calls directly into the
underlying zeroconf resolve_host() functions. As those return a list, we
get a TypeError("expected string or bytes-like object, got 'list'")>)
raised when trying to use the result of the lookup.

In https://github.com/esphome/esphome/pull/7747 the traceback was
fixed by just using the *first* element of that list, but the list
still wasn't *sorted*, so we're essentially using a *randomly*
selected address which happened to be the first one to respond to
mDNS.

Fix it properly by providing a helpers.sort_ip_addresses() function
which performs the same sort as we do for the OTA connection.

This involves calling gethostinfo() so we can see the address type
and scope for each address, then applying the same ordering criteria,
then getnameinfo() to turn it back into a string again for the output.

(In the original resolve_ip_address() we *wanted* the getaddrinfo()
tuples, so we didn't perform the final getnameinfo() and it all
seemed a little more natural).