[modbus_controller] Update state when ACK received for select, switch
Immediately update the state once an ACK has been received from the
device. This prevents waiting for an update of a component to see if
the update actually went through.

In turn this solves Home Assistant from showing new-state, old-state
and then new-state again.

Code is inspired by modbus_number, which already does this.