Fix BLE never connecting when connection takes longer than timeout + disconnect timeout
If the connection times out in aioesphomeapi, we send a disconnect request.
If the disconnect request times out, the connection is left in a pending
disconnect state. The retry happens but we are still in the connecting
state so we do not try to connect again. The connection eventually
gets made, but since we had a disconnect request pending we close
the connection right away, and aioesphomeapi sees this as it disconnected
right away and bleak-retry-connector tries again starting the cycle
over again.

Instead we now cancel the pending disconnect and let the connection
complete on the second attempt. This only happens if the device
takes a long time to establish the connection or the device
is at the edge of Bluetooth range.