Fix missing recursion guard release on ESP8266
When refactoring, I noticed that the recursion guard
was missing in the case we run out of space copying
the format string from flash.

This is a long standing issue that I noticed because
I was working in this area. It likely never happens
unless someone sets the tx_buffer to something very
low.

missing on `release` as well:
https://github.com/esphome/esphome/blob/7bb899bfa1caf116920a5d28e937f6b41659561c/esphome/components/logger/logger.cpp#L90