Revert "Convert to LockFreeQueue"
This reverts commit ba05cac936f9c3f57526425394098ca9184c54b8.

It was better off with the FreeRTOS queue. The consumer task was
hogging CPU precisely *because* of the lack of OS synchronization.

I'm also unconvinced that we really *did* have a single producer,
as MQTT messages might potentially come from anywhere, including log
messages from any thread. I'd rather be safe and future-proof.

Even when it was working, I saw repeated warnings of mqtt and other
tasks taking >30ms, which weren't happening with the previous version
that I'd had in production for months.