Attempt to fix "sleep" issue
If the light is off (brightness is 0) for some time, when switching it
on again via a single command which sets both brightness and color temp,
the color temperature is set too warm.

The fix idea is to set a different command every time brightness is 0,
namely a "turn on" command, right before actual brightness + color temp
is set.

Not tested yet, have some issues with platformio.