esp32_camera: deprecate i2c_pins; throw error if combined with i2c: block
If `i2c_pins:` is used in combination with an `i2c:` block, the i2c
module does not properly iterate to the next hardware. Leading
to both components trying to use the same hardware with different
GPIO pins.

This change will make it invalid to use `i2c_pins:` in combination
with an `i2c:` block and instead guide the user to use the new
`i2c_id:` option instead, which uses the i2c component by
reference.

In addition a warning will be printed, if `i2c_pins:` is used
reminding the existing userbase, that this option is now
deprecated (as it's no longer documented in the documentation).