[ush_host_hid] Implement USB HID support for US/ASCII keyboards
esp32 platforms s2, s3, p4 support USB **host** HID, meaning HID devices
such as keyboards and mouses can be connected and esp-idf can enumerate
and poll them.

This commit adds basic support for capturing keyboard key presses for a
standard ASCII (US) layout. It supports the 'shift' modifier for capital
letters.

I was not able to make this play nicely with the existing `usb_host`
component, because it also takes control of the devices when they
connect, which is incompatible with esp-usb's HID implementation.

The `check_class` tweak is necessary for weirdly-behaved devices which
act as a keyboard but do not report being keyboards, such as USB badge
readers.

This feature was successfully tested on an esp32-s3 devkit.

Naturally, this component plays nicely with a `key_collector`:

    usb_host_hid:
      keyboards:
        - id: kbd
          # ...

    key_collector:
      - id: reader
        source_id: kbd
        end_keys: "\r"