[switch) Add trigger `on_state_changed`
This adds a trigger `switch.on_state_changed` to the switch component, which will be triggered on any change to the switch (to `ON` or to `OFF`) and returns the new state (bool) to a var `x`.
It is very similar to the exiting `switch_on_turn_on` and `switch_on_turn_off`, and will even be trigger if any of those are triggered.
The new action gives more flexibility for dynamic actions that are common to both state changed, removing the need of code duplication.

The name `switch.on_state_changed` and not only `switch.on_state` was choosen to avoid misunderstanding based on `light.on_state`, which is triggered at the beginning of a light state transition, and not when the state was already changed (which is the case of `on_state_changed`).