Fix post-OTA logs display when using MQTT.
When transitioning to showing logs after an OTA, get_port_type() is
called with the device's hostname which returns "NETWORK" instead of
"MQTT".  This causes esphome to incorrectly claim no remote logging is
configured at the end of an OTA of an MQTT-only device, even though
running `esphome logs` afterwards works.

This change tweaks the logging logic to use MQTT logging in either
case when mqtt is available but native logs aren't, because `esphome
logs` still ends up passing in "MQTT" instead of the hostname to
show_logs().